Exclude ComputeTable from building
Set NDEBUG as CXXFLAGS to skip the compilation of this debug function.

The NDEBUG logic is broken because REGISTER_MODULE_INITIALIZER(snappy,
ComputeTable()); is set to always result to nothing, independently of
NDEBUG.

If debug tests should instead be included in the build, both NDEBUG
must not be set (as before) but also, REGISTER_MODULE_INITIALIZER
repaired.

COUCHDB-2463